test: refactor dts tests to split into multiple test files
