+++ tests/integration/dts/other-features/copy/copy.d.ts
