+++ tests/integration/dts/other-features/copy/package.json
