+++ tests/integration/dts/other-features/copy/rslib.config.ts
