+++ tests/integration/dts/other-features/copy/src/index.ts
