+++ tests/integration/dts/other-features/copy/tsconfig.json
